+++ src/commands/hardis/org/diagnose/legacyapi.ts
@@ -6,7 +6,7 @@ import * as c from "chalk";
 import * as sortArray from "sort-array";
 import { getCurrentGitBranch, isCI, uxLog } from "../../../../common/utils";
 import * as dns from "dns";
-import { canSendNotifications, getNotificationButtons, getOrgMarkdown, sendNotification } from "../../../../common/utils/notifUtils";
+import { canSendNotifications, getNotificationButtons, getOrgMarkdown, getSeverityIcon, sendNotification } from "../../../../common/utils/notifUtils";
 import { soqlQuery } from "../../../../common/utils/apiUtils";
 import { WebSocketClient } from "../../../../common/websocketClient";
 import { NotifProvider, NotifSeverity } from "../../../../common/notifProvider";
@@ -278,6 +278,9 @@ See article to solve issue before it's too late:
     const soonDeprecatedApiCalls = [];
     const endOfSupportApiCalls = [];
     const logEntries = await conn.request(logFileUrl);
+    const severityIconError = getSeverityIcon("error");
+    const severityIconWarning = getSeverityIcon("warning");
+    const severityIconInfo = getSeverityIcon("info");
     for (const logEntry of logEntries) {
       const apiVersion = logEntry.API_VERSION ? parseFloat(logEntry.API_VERSION) : parseFloat("999.0");
       // const apiType = logEntry.API_TYPE || null ;
@@ -292,11 +295,17 @@ See article to solve issue before it's too late:
           logEntry.SFDX_HARDIS_DEPRECATION_RELEASE = legacyApiDescriptor.deprecationRelease;
           logEntry.SFDX_HARDIS_SEVERITY = legacyApiDescriptor.severity;
           if (legacyApiDescriptor.severity === "ERROR") {
+            logEntry.severity = "error";
+            logEntry.severityIcon = severityIconError;
             deadApiCalls.push(logEntry);
           } else if (legacyApiDescriptor.severity === "WARNING") {
+            logEntry.severity = "warning";
+            logEntry.severityIcon = severityIconWarning;
             soonDeprecatedApiCalls.push(logEntry);
           } else {
             // severity === 'INFO'
+            logEntry.severity = "info";
+            logEntry.severityIcon = severityIconInfo;
             endOfSupportApiCalls.push(logEntry);
           }
           break;