+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 Note: Can be used with `sfdx plugins:install sfdx-hardis@beta` and docker image `hardisgroupcom/sfdx-hardis@beta`
 
+- Api logs enhancements:
+  - Add severity and severityIcon in all log elements details
+  - Add dateTime property (ISO format) in all API logs 
+
 ## [4.33.0] 2024-05-04
 
 - New notifications provider: **ApiProvider (beta)**, that allows to send notifications via HTTP/JSON to remote endpoints, like Grafana Loki, but also any custom one